fix: NotificationItem onApplySuggestion 파라미터 타입 변경

+++ android/app/src/main/java/com/mulkkam/ui/notification/component/NotificationItemComponent.kt
@@ -28,7 +28,7 @@ import java.time.LocalDateTime
 @Composable
 fun NotificationItemComponent(
     notification: Notification,
-    onApplySuggestion: (Int) -> Unit,
+    onApplySuggestion: () -> Unit,
     onRemove: () -> Unit,
     modifier: Modifier = Modifier,
 ) {
